Tremendously speedup reindexing by disabling sqlite synchronous mode
By default sqlite requests operating system to ensure that all content is safely written to the disk surface prior to continuing. If underlying filesystem is not mounted with nobarrier option and resides on rotary device flushing index.db to disk becomes bottleneck of full reindexing, speed might be about 3 pages per second.

Disabling sync makes reindexing 50-100 times faster is such cases.
Amount of written data on disk is reduced as well, which is desirable on SSD devices. On test notebook with 430 pages amount of written data during single reindex is reduced from 35MB to 300-400KB.

We don't need extra caution when dealing with cache. Even if gets corrupted (which is especially unlikely if we perform our operations fast) it may be regenerated any time.