Fallback to notebookdialog if default went missing - lp:1582330
Changed how notebook argument is used and made it robust for the
case where the default notebook is missing. User is presented with
the notebook list to allow changing the default.

Made it explicit in the code that only the two gui commands use
the default notebook, other commands require a notebook argument
on the commandline anyway.