versioncontrol improvements for git staging
Change stage() to mean "prepare the repo for a commit". It is called by SaveVersionDialog before getting the status of the repo. It is only implemented for git, where it does a 'git add' on the whole notebook, so that the status message accurately reflects what will be committed.

Remove '-a' from 'git commit'. Both stage() and VCSBackend.commit() call add() before commit(). Having '-a' here might capture unwanted changes in a zim commit, if the notebook is part of larger repo.

Remove use_staging. It was only used with git and the new way of staging means it isn't necessary.

Fixes #249