Fix rename page for case insensitive filesystem
Added test case to ensure all notebook API passes tests
as well for case insensitive filesystems.

In particular renaming to the same file but with different
case was broken due to "file.exists()" checks that can't
be trusted in this context.

Second order issue found was typo in SQL query when
resolving links which in turn led to placeholders for
the old page remaining.

Fixes #215 part 1