Do not suggest or create floating links for the user.
Floating links routinely become broken and misdirected during page renames.

Futhermore, the algorithim used to compute a floating link between two
pages has a high runtime cost, and does not scale well with a large
number of pages, particularly repos with many 'deep' pages.

This commit removes Zim's ability to generate floating links, while not
removing general support (both for data-in-place and as the user might
manually enter). Without any particular user interaction, links that
were originally created to be floating will gradually be replaced by
absolute links as their target pages are moved and renamed.

As a result of this commit, users might be happy to find considerably
faster "insert link" and "edit link" dialog box operations (specifically
with the building of the autocompletion suggestions), faster page move
and rename operations, and (in the long term) far fewer broken links.

A downside, to many users, might be a considerable increase in the
number of longer/absolute links. This might be ameliorated by using
the page title or 'n' of the last path segments as the visible link
text.

The generation of links between pages that do no share a parent are
practically not effected (differing only with a preceding colon ':'),
except for the fact that they are no longer vulnerable to the page
movement bug.