Remove notebook "profile"
Remove the option to assign notebooks a "profile" and have
preferences per profile. Any preference that a user would want
to switch per notebook or group of notebooks should be a
notebook property now. Having notebook properties for plugins
as well allows for this.

Also removing the "profile" breaks the dependency of the
preferences on the notebook, which simplifies some aspects
of the application.