Image generator plugins now are "inserted objects"
Removed most of the special handling for image generator plugins.
Instead these now define an InsertedObjectType which results in a
custom widget in the editor.

This did result in changed to the inserted objects API to allow
for all logic needed to render a file from the attachments folder.

For backward compatibility there is a fallback to rendering an
image in both the pageview and the export formats if the plugin
is not loaded.