Change behavior for lists with mixed bullets
Lists of mixed numbered and non-number bullets in the wiki text will
now be broken into seperate lists. Previous these were re-numbered
and forced into a single list type.

The editor also will no longer force renumbering of lists gets joined
due to the deletion of lines.

Bullet types are still converted when indenting of a list item is
changed, but checkboxes are treated as an exception to this rule.

Fixes #524