Add warnings if locale does not support unicode
Add warnings both in the terminal and the preferences dialog
if the preferred encoding does not support unicode. This will
lead to various errors that cannot be resolved. Users must start
zim with a locale encoding of UTF-8 or compatible in order to
use all functions.

Reverses commit 4a5770ac6b397ae43495dd814f87d91431798501
Unfixes #748