Exclude formatting when copying headings
Prior to this change, copying a heading in a notebook and pasting it
into another text editor would include the formatting. For example, the
level three heading "Title" would be pasted into a plain text document
as "### Title".

This is inconsistent with the copying and pasting of text with other
formatting, such as bold or italics. A bold word, such as "test", is
pasted as "test", not "**test**".

This change also affects the plain text export, meaning that headings
are not exported with any formatting.

Fixes #752
Closes #1082