Show plugins that failed to load in the list
This change addresses #668, which requires failed plugins to be
displayed in the list of plugins. This is only a minor improvement
which:
* includes a plugin that failed to load in the list
* displays a generic error message in the details pane
* uses the plugin's key (file or directory name) in the list if the
  plugin's proper name can't be read due to the load problem

The generic error message is also displayed at the top of the details
pane for plugins that can't be activated due to missing dependencies.
This may not be desired, in which case an additional field per plugin
model will be needed.