Show plugins that failed to load in the plugin list
This change addresses #668, which requires failed plugins to be
displayed in the list of plugins. This is only a minor improvement
which includes a plugin that failed to load in the list and
displays a generic error message in the details pane. It uses the
plugin's key (file or directory name) in the list if the
plugin's proper name can't be read due to the load problem

Closes #668
Closes #1086