Add dark mode preference to equation editor
If the user uses a dark mode, equations will be invisible because they
are rendered with a dark font color on a transparent background. The new
dark mode setting can be ticked in the plugin preferences of the
equation editor and will cause TeX to use a white font color for all
newly generated equations.

Fixes #1143