Add support for environ.ini at startup
Zim will now read a config file "environ.ini" at startup before
doing anything else. This file must be in the same folder as the
script or executable starting zim.

The rationale is to allow setting up zim for a portable installation
without the need to wrap the exe in a .bat file to modify the
environment.

Part of solution to issue #1206