Add extensions on PageWindow by splitting NotebookView and PageView
Introduces a "NotebookView" class and corresponding extension which
inherits from PageView. The MainWindow uses the NotebookView and the
PageWindow now uses the PageView and loads extensions. As a result
plugin extensions are active in page windows, but plugins can decide
on whether the extension is relevant for all page view or only for
the "notebook view".

Also updated PLUGIN_WRITING.md to explain this new extension class
and added a class diagram to visualize the plugin framework.