Add headerbar for fullscreen mode
Add a headerbar for fullscreen mode that is hidden untill the
mouse pointer touches the top of the screen. This allows access
to headerbar functions in fullscreen mode.
Based on UI design and using example code from GEdit.

Also fix CSS styling for the distractionfree plugin to
keep original background color for buttons in the editbar
and headerbar.