Interpret pasted text as "wiki" format and add "Paste As Verbatim"
Related to discussion of issue #1324 where it was pointed out that
when you paste text gets interpreted as format "plain" but upon
reloading still formats as wiki. Better to format as wiki directly
to avoid surprises. To allow pasting real plain text a
"Paste As Verbatim" is added to the context menu. Also pasting within
a verbatim formatted region will not format the text.