Fix bug with relative links in single page export
Changed the relative root of the single page exporter. It was
the attachement folder, but now it is the parent folder of the
main page. This allows linking up to the main page with relative
links as well. Potential side effect is to allow relative linking
other (external) files in the export folder as well. This is not
necessarily an issue and might even be used as a feature.

Fixes #473