Provide import option for non-page text files
If a text file is encountered in the notebook that is not a page
during "open page" or "new page" actions the user will be prompted
whether they want to import the text file as a page or not.

A "move page" action will fail with explanation when the target
location is blocked by a non-page text file.

Importing a non-page text file that is already part of the notebook
will convert the file in-place while importing a file from outside
the notebook will make a copy.

Closes #907