Add "classic" mode to Toolbar plugin
Add a "classic" mode to the Toolbar plugin which hides the pageview
edit bar and includes the edit tools in the toolbar instead. When
placed along the top of the window this restores the toolbar
behavior of previous versions. But now the implementation is an
optional plugin, which allows more flexibility.