Fixup of face0021e67de063e77cf2c6ac36d1e5915a63a4
Instead of using the notebook name for tracking links across
notebooks, use the interwiki key. And if such key is not available
convert the name into a valid key.

Without this fix the previous change broken copy pasting links
within notebooks where the name was not also a valid interwiki key