Try to use GtkSourceView 4.x if 3.x not found
This enables systems to use GtkSourceView 4.x in case they no longer
provide 3.x. This does not introduce changes for existing systems
as GtkSourceView 3.x is still the preferred version.

GtkSourceView 4.x is the last major version to support
Gtk 3 and requires at least 3.24, so it may not be suitable for older
distributions.