currently, images pasted from clipboard are inserted as-is, even if they do not fit to window/screen width.
this behavior makes it pretty hard to use ZIM as a notebook for documenting online meetings, as capturing screenshots is a common task that often helps to caputure screen content presented in online sessions.

this commit scales down pasted images to a max width, enabling the user to see the full pasted picture that can then more easily be integrated into a document.

I would like to add an option in the notebook settings dialog to let the user define the desired setting for the max width value.
Any hint on how to get this done in the source code is appreciated.