Also match URL in wiki text for local network
In the GFM spec a URL most contain 2 or more domain fragments to
be recognized as an URL and auto-linked. In this patch that
restriction is removed to also let localhost and local network
links match. As pointed out on the mailing list, this makes sense
for a desktop wiki compared to an online wiki.