Add confirmation dialog when trashing pages
Since pages are moved to system trash on "delete page" and cannot
be restored from within zim this needs a confirmation dialog after
all. Apart from confirmation on "destructive" action the purpose
of the dialog is also to inform users that the system trash is used.

Considered making this action undoable by implementing a local trash
but that seems to heavy since any deletion with "<Ctrl>A <Delete>"
is not captured and also the updating of links is not easily reverted.
Therefore version control is the better option for throughly restoring
pages.

Closes #1038