Make indexer robust against false folder mtime
Make the file indexer rebust against filesystems that do not
properly update the mtime on folders when files are created
or deleted. Several cases have been seen where this happens
with cloud synchronized folders or folders on secure drives.

Could have a penalty on indexing speed, but no significant
impact seen when tested against the notebook generated by
"./tools/create_large_test_notebook.py".

Also updated the behavior of "zim --index" to match the behavior
of the menu action "Check and Update Index". Added a "--flush"
or "-f" switch to force the old behavior of "--index" and do
a full re-build of the index without checking.

Closes #1481