Add debug and plugin info about "babel" library
The documentation for the Journal plugin mentions that Babel is an optional Python library to make weekdays work based on locale.

This adds mentioning that optional dependency in the Journal plugin information dialog, including a check if it is installed.

It also adds a debug message if there is an ImportError when trying to load Babel.