Minimal support for "Terminal" option in desktop files
Since there is no canonical way to know what terminal to call (and how to call it)
this is only minimal support if "Terminal" is set to "true". No intention to
make is support all possible platforms and enviroments.

Still recommend to just create a ".desktop" entry which explicitly states the
terminal emulator you want to use in the "Exec" field.

Closes #1931