Refactor preference for custom window decoration
This change introduces a global preference for using
custom window decoration instead of a classic toolbar.
On Windows and Mac OS the default is off and the classic
toolbar is used with os specific window decoration.

This reduces the function of the "toolbar" plugin to
customizing the classic toolbar.

The tasklist plugin also respects the same global
preference.

Fixes #1983
Closes #2007