Use Homebrew to run test suite on macOS
Remove the dependency on https://gitlab.com/dehesselle/zim_macos.
While this introduces a small risk as other libraries and versions
are used to run the tests than the ones that make up Zim.app, it
makes the CI standalone again.

Disable TestPlugins() test case as it segfaults for yet unknown
reasons on macOS.