Avoid incompatible screenshot utils on Wayland
Previously, the Insert Screenshot plugin chose the list of allowed
screenshot utilities, as well as the default option, based on
the platform, either 'Linux', 'Windows' or 'Darwin'. However, most
choices for 'Linux', including the default utility 'import' were not
compatible with Wayland. Improve platform detection to detect the
Wayland case separately and offer only compatible utilities. This means
that at least for now, 'gnome-screenshot' is the only option on Wayland.