Store the returned pid until callback
On Windows, the returned PID from GObject.spawn_async is a
handle, which ownership is transferred to the caller.

If this handle goes out of scope, it is freed, and its subsequent usage
in WaitForMultipleObjectsEx through GObject.child_watch_add results in
repeated failed calls and high CPU usage.

Fixes #1697