Avoid unhandled exceptions in `NotebookOperation`
`NotebookOperation` is intended to be executed in the GTK main loop via
`run_on_idle()`. Even if such execution terminates in an unhandled
exception, execution of Zim still continues. Thus, normally the effect
of an unhandled exception is to print an error message and a traceback
in Zim's debug log.

However, if `sys.excepthook` is overridden, different action may
result. In particular, in Fedora Linux, Automatic Bug Reporting Tool
(ABRT)'s Python plugin overrides `sys.excepthook` to notify the user
about unhandled exception in Zim, and also to upload a crash report to
Fedora Analytics Framework. Since multiple `NotebookOperation`
subclasses use exceptions to communicate failures that are preventable
by the user, such as trying to create a page that already exists, this
is not desirable.

To avoid triggering `sys.excepthook`, stop re-raising exceptions in
the already existing `NotebookOperation` catch-all exception handler.
Just print them to the debug log instead.

Note that since the catch-all handler already communicates
the exceptions to `NotebookOperation` caller via the `cancelled` and
`exception` properties, it could be argued that all handling of
exceptions, even logging them, should be left to the caller. However,
there are places where `NotebookOperation` subclasses are executed, but
those properties are not checked afterwards at all, and thus the *only*
error handling they already have is the debug log. Since the author of
this change does not feel comfortable adding exception handling to all
those places, the solution is to keep logging the exception inside
`NotebookOperation`.