Fix occurence of metadata in parse tree during page indexing
By default, the wiki format parser does not strip the header of a wiki page.
Hence, when reading a page from disk without specifying file_input=True as
optional parameter, the metadata will be part of the resulting parsetree
just like normal text.

This fix is needed so that the full-text indexing plugin does not index
the metadata as normal text.