Sort search results in GUI and CLI first by score, then by name
This is achieved by using the stable sort properties of python's sort
function which doesn't change the order of entries with the same sort
value.

The GUI search already produced correctly sorted results for the
traditional search by accident (I think), but this was messed up
a bit by the indexed full-text search plugin. This patch fixed the sort
order explicitly.