Add Indexed Full-Text Search plugin without search integration
This plugin only brings in functionality for populating the full-text index
using sqlite's FTS5 module.

The current implementation assumes that after enabling, all changes to the
index are made with the plugin enabled. Any change to the notebook with the
plugin disabled will lead to an out-of-sync index for full-text search,
which currently can only be recovered by regenerating the index.