Add preference to toggle dark theme variant
Add a preference which requests Gtk to give zim a dark variant
of the active theme. For themes which support this.

In addition add support in `style.conf` to give both a normal
color and a `darktheme` variant of the same color so style
is modified for dark themes accordingly.

And make the colors used in the tasklist configurable as
well via `style.conf`.

Closes #1727