Fix: remove pageindex last selected filter
The "last selected" filter is unneccessare because there is no event triggered if the page is already selected.

Also this filter causes a bug where selecting the previous selected page does not work, if the page switching happened with anything else then selecting it in the pageindex.
E.g. clicking a link in a page to a page, or using the back-button.