Add keybind for ToC plugin float size toggle
The tableofcontents plugin has a mode in which the table floats above
the page, and where clicking its title toggles minimised/expanded size.
This is very useful, since the table obscures text on the page, but is
inconvenient to access with the mouse.
Add an accelerator action (defaults to `<ctrl>grave`) to be able to
toggle the size via keyboard.